Added welcome.jsp and made changes to package
