+++ ISDAssignment1/nbproject/private/private.properties
@@ -1,10 +1,10 @@
-deploy.ant.properties.file=C:\\Users\\notba\\AppData\\Roaming\\NetBeans\\12.6\\config\\GlassFishEE6\\Properties\\gfv3-1458661859.properties
+deploy.ant.properties.file=/Users/pyaephyozaw/Library/Application Support/NetBeans/12.4/config/GlassFishEE6/Properties/gfv3-139759611.properties
 j2ee.platform.is.jsr109=true
-j2ee.server.domain=C:/Users/notba/GlassFish_Server/glassfish/domains/domain1
-j2ee.server.home=C:/Users/notba/GlassFish_Server/glassfish
-j2ee.server.instance=[C:\\Users\\notba\\GlassFish_Server\\glassfish;C:\\Users\\notba\\GlassFish_Server\\glassfish\\domains\\domain1]deployer:gfv3ee6wc:localhost:4848
-j2ee.server.middleware=C:/Users/notba/GlassFish_Server
+j2ee.server.domain=/Users/pyaephyozaw/GlassFish_Server/glassfish/domains/domain1
+j2ee.server.home=/Users/pyaephyozaw/GlassFish_Server/glassfish
+j2ee.server.instance=[/Users/pyaephyozaw/GlassFish_Server/glassfish:/Users/pyaephyozaw/GlassFish_Server/glassfish/domains/domain1]deployer:gfv3ee6wc:localhost:4848
+j2ee.server.middleware=/Users/pyaephyozaw/GlassFish_Server
 javac.debug=true
 javadoc.preview=true
 selected.browser=SL[/Browsers/FirefoxBrowser
-user.properties.file=C:\\Users\\notba\\AppData\\Roaming\\NetBeans\\12.6\\build.properties
+user.properties.file=/Users/pyaephyozaw/Library/Application Support/NetBeans/12.4/build.properties
