+++ ISDAssignment1/src/java/uts/isd/model/Product.java
@@ -1,4 +1,4 @@
-package User;
+package uts.isd.model;
 import java.io.Serializable;
