+++ ISDAssignment1/web/welcome.jsp
@@ -1,55 +1,43 @@
 <%-- 
-    Document   : welcome
-    Created on : 29/03/2024, 2:47:44 PM
-    Author     : notba
+    Document   : main
+    Created on : Mar 28, 2024, 3:29:37 PM
+    Author     : pyaephyozaw
 --%>
 
 <%@page contentType="text/html" pageEncoding="UTF-8"%>
-<%@page import="User.*"%>
+<%@page import="uts.isd.model.*"%>
 <!DOCTYPE html>
 <html>
     <head>
         <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-        <title>Welcome-Kat</title>
+        <link rel="stylesheet" href="css/layout.css">
+        <title>Welcome Page</title>
     </head>
     <body>
+        <%           
+          User user = (User) session.getAttribute("user");
+        %>
         
-      <%User user = (User)session.getAttribute("user");
-      String submitted="nothing";
-      submitted = request.getParameter("submitted");
-      %>
-      
-      
-      <div class="menu">
+        <div class="menu">
             <ul>
-                <li><a href="http://localhost:8080/ISDAssignment1/">Home</a></li>
-                <li><a href="http://localhost:8080/ISDAssignment1/login.jsp" > Login</a></li>
-                <li><a href="http://localhost:8080/ISDAssignment1/register.jsp" > Register</a></li>
-                <li><a href="http://localhost:8080/ISDAssignment1/welcome.jsp" > You</a></li>
+                <li><a id="active" href="index.jsp">Home</a></li>
+                <li><a href="logout.jsp">Logout</a></li>
             </ul>
-            </div>
-            <br>
-            <br>
-            
-            <% if (user==null){ %>
-            <h2>Hi anonymous user</h2>
-            <h2>Creating an Account?</h2>
-            <button><a href="http://localhost:8080/ISDAssignment1/register.jsp">Register</a></button>
-            
-            <% } %>
-      
-      
-     
-      <% if ((user!=null) && (submitted).equals("true")){
-                user.setName(request.getParameter("name"));
-                user.setEmail(request.getParameter("email"));
-                user.setPhone(request.getParameter("phonenumber"));
-                user.setAddress(request.getParameter("address"));
-       } %>
-               
-                User's name is <%=user.getName()%>
-               User's email is <%=user.getEmail()%>
-               User's phone number is <%=user.getPhone()%>
-               User's address is <%=user.getAddress()%>
+        </div>
+        <br>  
+        <br>
+        <div>
+        <%if(user != null) {%>
+        <h1>Hello <%=user.getName()%>,
+            <br> 
+            Welcome to IoTBay Web Application
+        </h1>
+        <% } else {%>
+        <h1> Hello Anonymous User, </h1>
+        <h2> Do you want to register? </h2>
+        <button><a href="register.jsp">Register</a></button>
+        
+        <% } %>
+        </div>
     </body>
 </html>
